move formats to submodule (#236)
