+++ src/canmatrix/formats/fibex.py
