+++ src/canmatrix/formats/sym.py
