+++ src/canmatrix/tests/test_arxml.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
-import canmatrix.arxml
+import canmatrix.formats.arxml
 import pathlib2
 
 
 def test_ecu_extract():
     here = pathlib2.Path(__file__).parent
     
-    db = canmatrix.arxml.load(str(here / "MyECU.ecuc.arxml"))['']
+    db = canmatrix.formats.arxml.load(str(here / "MyECU.ecuc.arxml"))['']
     assert db.frames is not None
     assert len(db.frames) == 2
     assert len(db.frames[0].signals) == 3
